Safely update the hostname entries in hosts file.
I was thinking that this code might cause duplicate entries in the hosts file if the hostname already exists. This might be bad if the IP addresses differ. So I added some logic to try and do a find and replace first, and if there is no existing hostname then we will append it to the hosts file.